Inject util in context
